added e2e test of gcp broker with four sources and fixed typo of cloudschedulersource example
