+++ test/e2e/e2e_test.go
@@ -313,3 +313,31 @@ func TestGCPBroker(t *testing.T) {
 	defer cancel()
 	GCPBrokerTestImpl(t, authConfig)
 }
+
+// TestCloudPubSubSourceWithGCPBroker tests we can knock a Knative Service from a GCPBroker from a CloudPubSubSource.
+func TestCloudPubSubSourceWithGCPBroker(t *testing.T) {
+	cancel := logstream.Start(t)
+	defer cancel()
+	PubSubSourceWithGCPBrokerTestImpl(t, authConfig)
+}
+
+// TestCloudStorageSourceWithGCPBroker tests we can knock a Knative Service from a GCPBroker from a CloudStorageSource.
+func TestCloudStorageSourceWithGCPBroker(t *testing.T) {
+	cancel := logstream.Start(t)
+	defer cancel()
+	StorageSourceWithGCPBrokerTestImpl(t, authConfig)
+}
+
+// TestCloudAuditLogsSourceWithGCPBroker tests we can knock a Knative Service from a GCPBroker from a CloudAuditLogsSource.
+func TestCloudAuditLogsSourceWithGCPBroker(t *testing.T) {
+	cancel := logstream.Start(t)
+	defer cancel()
+	AuditLogsSourceBrokerWithGCPBrokerTestImpl(t, authConfig)
+}
+
+// TestCloudSchedulerSourceWithGCPBroker tests we can knock a Knative Service from a GCPBroker from a CloudSchedulerSource.
+func TestCloudSchedulerSourceWithGCPBroker(t *testing.T) {
+	cancel := logstream.Start(t)
+	defer cancel()
+	SchedulerSourceWithGCPBrokerTestImpl(t, authConfig)
+}
