+++ test/e2e/test_auditlogs.go
@@ -31,11 +31,6 @@ import (
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 )
 
-const (
-	serviceName = "pubsub.googleapis.com"
-	methodName  = "google.pubsub.v1.Publisher.CreateTopic"
-)
-
 func CloudAuditLogsSourceWithTestImpl(t *testing.T, authConfig lib.AuthConfig) {
 	project := os.Getenv(lib.ProwProjectKey)
 
@@ -48,15 +43,15 @@ func CloudAuditLogsSourceWithTestImpl(t *testing.T, authConfig lib.AuthConfig) {
 	defer lib.TearDown(client)
 
 	// Create a target Job to receive the events.
-	lib.MakeAuditLogsJobOrDie(client, methodName, project, resourceName, serviceName, targetName)
+	lib.MakeAuditLogsJobOrDie(client, lib.MethodName, project, resourceName, lib.ServiceName, targetName)
 
 	// Create the CloudAuditLogsSource.
 	lib.MakeAuditLogsOrDie(client,
 		auditlogsName,
-		methodName,
+		lib.MethodName,
 		project,
 		resourceName,
-		serviceName,
+		lib.ServiceName,
 		targetName,
 		authConfig.PubsubServiceAccount,
 	)
